+++ src/main/java/net/alvin/infinityforge/infinity/abilities/impl/reality/SpawnItemAbility.java
@@ -1,10 +1,11 @@
 package net.alvin.infinityforge.infinity.abilities.impl.reality;
 
 import net.alvin.infinityforge.infinity.abilities.base.AbilityIcon;
-import net.alvin.infinityforge.infinity.abilities.base.ActiveAbility;
 import net.alvin.infinityforge.infinity.InfinityStoneType;
+import net.alvin.infinityforge.infinity.abilities.ext.StatefulAbility;
 import net.alvin.infinityforge.item.FakeItem;
 import net.minecraft.entity.ItemEntity;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -14,7 +15,7 @@ import net.minecraft.util.Identifier;
 import java.util.List;
 import java.util.function.Supplier;
 
-public class SpawnItemAbility extends ActiveAbility {
+public class SpawnItemAbility extends StatefulAbility<Item> {
     private final boolean spawnFake;
 
     public SpawnItemAbility(Identifier id, AbilityIcon icon,
@@ -28,8 +29,19 @@ public class SpawnItemAbility extends ActiveAbility {
     @Override
     public boolean onActivate(ServerWorld world, ServerPlayerEntity player, List<InfinityStoneType> activeStones) {
         ItemStack fakeItem = FakeItem.create(Items.DIAMOND, 64);
+        setState(player, FakeItem.getDisguise(fakeItem));
         ItemEntity entity = new ItemEntity(world, player.getX(), player.getY(), player.getZ() + 1.0, fakeItem);
         world.spawnEntity(entity);
         return false;
+    }
+
+    @Override
+    protected ItemStack getStateIconStack(Item state) {
+        return spawnFake ? FakeItem.create(state, 1) : new ItemStack(state);
+    }
+
+    @Override
+    protected Class<Item> getStateType() {
+        return Item.class;
     }
 }