+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New AbilityIcon system.
 - Some new icons.
 - Added some config options.
+- Added Stone World Gen.
+- Internal changes for performance.
+- Added dynamic ability icons.
 
 ## [0.6.0-beta.1] - 2026-05-20